chore: rename wallet-networks-config package to wallet-config
