+++ packages/suite/package.json
@@ -38,7 +38,7 @@
         "@trezor/validation": "*",
         "@suite-common/intl-types": "*",
         "@suite-common/metadata-types": "*",
-        "@suite-common/wallet-networks-config": "*",
+        "@suite-common/wallet-config": "*",
         "@suite-common/wallet-types": "*",
         "@suite-common/wallet-constants": "*",
         "bignumber.js": "^9.0.2",
