+++ packages/suite/src/config/wallet/index.ts
@@ -1,4 +1,4 @@
-import { networksCompatibility } from '@suite-common/wallet-networks-config';
+import { networksCompatibility } from '@suite-common/wallet-config';
 
-// TODO: remove this reexport and import directly from @suite-common/wallet-networks-config
+// TODO: remove this reexport and import directly from @suite-common/wallet-config
 export { networksCompatibility as NETWORKS };
