+++ suite-common/wallet-config/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@suite-common/wallet-networks-config",
+    "name": "@suite-common/wallet-config",
     "version": "1.0.0",
     "private": true,
     "license": "See LICENSE.md in repo root",
