+++ suite-common/wallet-config/src/index.ts
