+++ suite-common/wallet-config/src/networksConfig.ts
