+++ suite-common/wallet-config/tsconfig.json
