+++ suite-common/wallet-types/package.json
@@ -10,7 +10,7 @@
         "type-check": "tsc --build"
     },
     "dependencies": {
-        "@suite-common/wallet-networks-config": "*",
+        "@suite-common/wallet-config": "*",
         "@suite-common/wallet-constants": "*",
         "@suite-common/intl-types": "*",
         "@trezor/connect": "9.0.0-beta.4",
